backend/app/services/pdf_parser.py
import PyPDF2
import io

async def parse_pdf_pypdf(file_content: bytes) -> str:
    """
    Parses a PDF file using PyPDF2 and extracts text content.
    """
    text = ""
    try:
        pdf_file = io.BytesIO(file_content)
        pdf_reader = PyPDF2.PdfReader(pdf_file)
        for page_num in range(len(pdf_reader.pages)):
            page = pdf_reader.pages[page_num]
            text += page.extract_text() or ""  # Add empty string if None
    except Exception as e:
        # Handle parsing errors, e.g., log them
        print(f"Error parsing PDF with PyPDF: {e}")
        # Optionally, re-raise or return an error indicator
        raise ValueError(f"Failed to parse PDF with PyPDF: {e}") from e
    return text

# Placeholder for Gemini parsing
async def parse_pdf_gemini(file_content: bytes, api_key: str) -> str:
    # This will be implemented later
    pass

# Placeholder for summarization
async def summarize_text_gemini(text: str, api_key: str) -> str:
    # This will be implemented later
    pass 